Change preeval behavior and improve template variables
This commit changes the behavior of preeval to only affect when a child
template is added as a component to a TemplateGenerator. This means a
non-preeval TemplateGenerator can still be evaluated in memory via
evaluate(). Preeval just causes a preevaluation for when the template is
added.

Also, the behavior of template variables is changed slightly to improve
variable propogation and to avoid redundant calls to htmlentities().
Namely, the default and local variables were separated out to be escaped
separately, and flags were added to annotate the status of escaped
variables. Also, variables from parent TemplateGenerators are only added
to the local list that is extracted as opposed to the list bound to the
object. In this way there is only one copy of the variable that needs to
be escaped before being added to the local execution scope.